Add support for multiple users and protect todo endpoints
CORS is still giving me a hard time, but I did manage a running config for now.

+++ src/main/java/br/com/klimber/inova/endpoint/TodoEndpoint.java
@@ -2,10 +2,13 @@ package br.com.klimber.inova.endpoint;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.common.exceptions.UnauthorizedUserException;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -16,6 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
+import br.com.klimber.inova.model.Customer;
 import br.com.klimber.inova.model.Todo;
 import br.com.klimber.inova.repository.TodoRepository;
 
@@ -27,38 +31,56 @@ public class TodoEndpoint {
 
 	@GetMapping("/todoApp/todos")
 	public List<Todo> getTodos() {
-		return todoRepository.findAll();
+		Customer customer = (Customer) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		return todoRepository.findByCustomer(customer);
 	}
 
 	@PostMapping("/todoApp/todos")
 	public Todo addTodo(@RequestBody Todo todo) {
+		Customer customer = (Customer) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		todo.setCustomer(customer);
 		return todoRepository.save(todo);
 	}
 
 	@PatchMapping("/todoApp/todos/{id}")
 	public Todo updateTodo(@PathVariable("id") Integer id, @RequestBody Todo todo) {
+		Customer customer = (Customer) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		Optional<Todo> currentTodo = todoRepository.findById(id);
+		if (currentTodo.isEmpty() || currentTodo.get().getCustomer().getId() != customer.getId()) {
+			throw new UnauthorizedUserException("Todo does not exist or is not owned by user");
+		}
 		todo.setId(id);
+		todo.setCustomer(customer);
 		return todoRepository.save(todo);
 	}
 
 	@DeleteMapping("/todoApp/todos/{id}")
-	public String deleteTodo(@PathVariable Integer id) {
+	public ResponseEntity<Object> deleteTodo(@PathVariable Integer id) {
+		Customer customer = (Customer) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		Optional<Todo> currentTodo = todoRepository.findById(id);
+		if (currentTodo.isEmpty() || currentTodo.get().getCustomer().getId() != customer.getId()) {
+			throw new UnauthorizedUserException("Todo does not exist or is not owned by user");
+		}
 		todoRepository.deleteById(id);
-		return "Success";
+		return new ResponseEntity<Object>(HttpStatus.OK);
 	}
 
 	@PatchMapping("/todoApp/checkAllTodos")
-	public String checkAll(@RequestBody JsonNode completed) {
-		List<Todo> allTodos = todoRepository.findAll();
+	public ResponseEntity<Object> checkAll(@RequestBody JsonNode completed) {
+		Customer customer = (Customer) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		List<Todo> allTodos = todoRepository.findByCustomerAndCompleted(customer,
+				!completed.get("completed").asBoolean());
 		allTodos.forEach(todo -> todo.setCompleted(completed.get("completed").asBoolean()));
 		todoRepository.saveAll(allTodos);
-		return "Success";
+		return new ResponseEntity<Object>(HttpStatus.OK);
 	}
 
 	@DeleteMapping("/todoApp/todosClearCompleted")
-	public String todosClearCompleted(@RequestBody JsonNode data) {
-		data.get("todos").forEach(id -> todoRepository.deleteById(id.asInt()));
-		return "Success";
+	public ResponseEntity<Object> todosClearCompleted() {
+		Customer customer = (Customer) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		List<Todo> completedTodos = todoRepository.findByCustomerAndCompleted(customer, true);
+		todoRepository.deleteAll(completedTodos);
+		return new ResponseEntity<Object>(HttpStatus.OK);
 	}
 
 	/**
@@ -67,8 +89,9 @@ public class TodoEndpoint {
 	@GetMapping("/todoApp/initTodos")
 	public void initTodos() {
 		List<Todo> listaInit = new ArrayList<>();
-		listaInit.add(new Todo("From Database", false));
-		listaInit.add(new Todo("Another One", false));
+		Customer customer = Customer.builder().id(1L).build();
+		listaInit.add(new Todo("From Database", false, customer));
+		listaInit.add(new Todo("Another One", false, customer));
 		todoRepository.saveAll(listaInit);
 	}
